add deleteProduct ADMINUI, deleteSale ADMINUI

+++ AdminUI/AdminUI/src/app/sale/sale.page.ts
@@ -84,6 +84,8 @@ export class SalePage implements OnInit {
     })
   }
   deletesale(s:IVendingMachineSale) {
+    if(! confirm('Are you sure?'))return;
+
     this.apiService.deleteSale(s.id).subscribe(rx => {
       console.log(rx);
       if (rx.status) {
